refactor: LangevinDynamics
